Create a seperate Package for friendsd
git-svn-id: https://gpsdrive.svn.sourceforge.net/svnroot/gpsdrive/trunk@2339 e3ca9d6f-f84a-410d-b97a-e7ddc331358d

+++ debian/move_utils.sh
@@ -7,8 +7,14 @@ destdir=debian/gpsdrive
 destdir_utils=debian/gpsdrive-utils
 mkdir -p $destdir_utils
 
+for file in `find ${destdir}/usr/local/ -type f` ; do
+    dst_name="${destdir_utils}/${file#${destdir}}"
+    dir=`dirname $dst_name`
+    mkdir -p $dir
+    mv -f "${file}" "${dst_name}"
+done
+
 mkdir -p "${destdir_utils}/usr/"
-mv -f ${destdir}/usr/local ${destdir_utils}/usr/
 mkdir -p ${destdir_utils}/usr/bin
 mv -f ${destdir}/usr/bin/*.pl ${destdir_utils}/usr/bin
 mv -f ${destdir}/usr/bin/*.py ${destdir_utils}/usr/bin
@@ -19,5 +25,6 @@ for file in ${destdir_utils}/usr/bin/*; do
     filename=${filename%.sh}
     filename=${filename%.pl}
     filename=${filename%.py}
+    test -e "${destdir}/usr/share/man/man1/$filename.1" || continue
     mv -f ${destdir}/usr/share/man/man1/$filename.1  ${destdir_utils}/usr/share/man/man1/
 done
